// src/app/interfaces/history.interface.ts
export interface History {
  _id: string;
  dateRequest: string;
  contractor: string;
  authorizationCode: string;
  policy: string;
  document: string;
  name: string;
  lastName: string;
  phone: string;
  age: number;
  typeAge: boolean;
  gender: number;
  address: string;
  reference: string;
  diagnostic: string;
  symptoms: string;
  treatment: string;
  medic: string;
  driver: string;
  isActive: boolean;
}